Include 64-bit StepManiaX DLL for Windows
This DLL was compiled for 64-bit Windows using VS2022. See https://github.com/sukibaby/stepmaniax-sdk for the 64 bit build used. This DLL is to be used with the new StepManiaX driver for Windows. The new CMake code will copy the DLL from the Win32 archutils directory to the Programs directory.