Use std::pair for RageTimers.
The old C++98 code had m_secs and m_us as separate member variables which was not ideal. This change encapsulates the two member variables within a std::pair into a single logical unit (m_time).

It was necessary to add getter and setter methods to maintain encapsulation and allow the code in Threads_Pthreads to work as intended for Mac/Unix.