Rewrite AddVertices(int Add)
This function is somewhat cryptic and does not seem well-optimized. The parameter is passed an an int instead of a size_t because it is a value that comes from/goes to Lua, and Lua sends/recieves int's. Instead of storing a size_t value in an int, we make sure the int is non-negative and then static_cast it to size_t when it's needed. A range-based loop is then used to make sure all elements of AMV_Tweens are resized.