Prepare for std::round: static_cast and round value which is already clamped to the range of [0,255]
This is surprisingly tricky: this operation is actually very expensive, taking up to 5ms on its own, and prone to failure on MSVC due to the strictness of what values are accepted by std::clamp. Add a new function "ClampToByte" to handle our needs more gracefully. This brings the scaling function down to about 1.0-1.5ms.

https://godbolt.org/z/63oWjKMvj