Hash-based ReloadFromSongDir
Currently, when the user selects a song, the song is loaded in from disk into a std::map. This std::map is compared to a std::map of the cached song. If any differences are detected, they are copied on top of the original pointers. In the end, the user gets some mix of the cached song plus the file on disk.

Now, the hash is calculated for the cached song and the file on disk. We return early if the hashes match, leaving pointers intact and untouched. If the file on disk has a different hash than the cached file, we remove the cached data and reload from disk into the cache before the song starts.

The new code is also greatly simplified by the use of range-based for loops instead of traditional for loops. This allows us to remove complex manual iterator handling logic.

I have removed the group offset logic is removed from ReloadFromSongDir because it is already handled when necessary by the LoadFromSongDir(sDir) call.

This commit can also guarantee that the chart the user plays has a verified hash, and corresponds to a known SHA-1.

Unintended side effect: In Edit mode, "Reload from Cache" is now functionally identical to "Reload from Disk", because in either case the hash will be found to differ and the game will reload from disk. The behavior of "Reload from Cache" may want to be evaluated.