Stop playing non-active videos more aggressively.
Currently, if you peek at the status of any given video texture, it will continue playing until it's finished, even if the status is a negative number. This results in a lot of videos continuing to loop unnecessarily. This can detect if a video is stuck in a failure state and  break out of the loop sooner. last_status should be thread_local to prevent thread contention on this variable or inaccurate statuses due to multi threaded access.