GraphicsWindow: refactor DestroyGraphicsWindow()
On Windows, when using fullscreen mode, there will be an ugly display of the flicker when exiting the game as the OS tries to restore its video settings. Here we reset the display mode to the monitor that was in fullscreen -before- destroying the window, which leads to a smooth transition when closing the game.

It also removes the nullptr checks, because we want to call the relevant release or destroy methods if the window or its resources exist at all.