GraphicsWindow: restore display mode before destroying window
On Windows, if fullscreen mode is used, there is an ugly display flicker when the game is quit as the OS tries to restore its display mode. This change will properly reset the display mode for the monitor which was in fullscreen mode before destroying the window and its resources. This change ensures a smooth, flicker-free transition as the game is exited.