Only use string literals when logging.
The main change here is to bring string construction out of the log itself, and change Windows specific functions werr_ssprintf and hr_ssprintf to not have an internal ssprintf call. Now we construct the string separately, and then pass the constructed string to the Log or Throw call for added safety. Additionally some braces were added where they were missing. Additionally note in DSoundHelpers I changed which variable is being statically cast, to make it more clear to the reader that this is never intended to be a negative number, since previously the DWORD was cast to int.