RageSoundReader_MP3: refactor id3_tag_query
This modular approach correctly handles more edge cases than the existing approach, and more self-documenting, helping to make the libmad implementation less opaque. As far as I can tell, this code is capable of accurately detecting the version and the exact size of the ID3 tag, even in files with mixed version ID3 tags.