RageSoundReader_MP3: remove leftover debugging log
This log is left over from my own debugging when I was doing work in this file to fix the accuracy of the libmad interface. It's often expected behavior that we call SetPosition_estimate (for example while a song preview is looping from an MP3 source) so this is effectively just noise in the logfile. When we call SetPosition_estimate, it is based on accurate data, and will be followed up by a SetPosition_hard call regardless.