WaveOut: refactor buffer size calculation, reorganize variables
We will now target a specific latency based on sample rate, which is the best approach for WaveOut when dealing with multiple possible target sample rates with the goal of minimizing impact on the WaveOut buffer latency, as this will in turn affect the global offset. Variable names are changed here, since the constants should not be MACRO_CASE, and many have been rearranged for this commit.