First attempt at fixing MovieTexture issues
Theory: the decoder thread stores its state in member variables, but there's a lack of synchronization between the decoding thread and main thread. Begin by implementing an atomic total_frames variable. This does require some local variables to be used for things like logging of this value.  We also introduce a singular state mutex to MovieTexture Generic, and put the check for the failure_ variable inside a lock controlled by the state mutex.