Break out string comparison to helper function
Implement a garbage collector for Lua

It will perform garbage collection upon loading ScreenSelectMusic, because we know there will be at least a moment where no time critical operations are happening, and we definitely don't have to perform garbage collection while a song is active.

Clear the stack in LuaManager::Release

Objects on the stack can not be garbage collected, so this exists to prevent unexpected behavior or memory leaks.

Revert "Implement a garbage collector for Lua"

This reverts commit 75896b125f7a7aa798bd7c89bd8af9118f8ea571.

Revert "Clear the stack in LuaManager::Release"

This reverts commit 68741b854a44d71037534efe563d7fb0ae8a6069.