Fix resampling mistake in RageSoundReader_Merge
The original code is modifying a copy of pointer. Ultimately, it does affect the actual elements stored in m_aSounds. The replacement code is more concise and more correct, by ensuring that the correct element of m_aSounds is resampled. This change is present in the aaa2 branch that made the recently released unofficial beta.