RageSoundReader_Merge correctness
Refactoring range-based for loops with index-based loops using size_t for correctness and explicitness.   In particular, fixes a 12 year old bug relating to RageSoundReader_Resample_Good being called on a copy of the element as opposed to the element of m_aSounds, a major bug which prevented resampling from working as expected.