Finalize Unix implementation
Many thanks are due to @sergioperez who put in a lot of effort helping to determine the most optimal way to handle ALSA-only systems. It seems like ALSA often negotiates a 48000 Hz sample rate with the hardware if one initializes an ALSA driver at 44100 Hz or below, so we'll fall back to 48000 if we're on Unix and don't have PulseAudio available.