! Parity with 12-07-25's test build
Because the sample rate is locked in before the audio driver is initialized, it's known and constant for the lifetime of the program, so there's no need to query the preferences manager for this value. We can simply query the global variable. Also, the global variable is now stored in the global header, so that we don't need to include RageSound's header to give access to the fallback variable, as it's used in many different places.