Placeholder for unfinished PulseAudio sample rate detection
The console based method of asking pactl what the sample rate is works, but is hideous.  I've set this up so there is already a framework to conditionally behave differently depending on whether or not the game was compiled with PulseAudio support. I just haven't implemented the PulseAudio sample rate detection via the API.