RageBitmapTexture: Cache maximum texture size
Every time we load a texture, we ask the video driver what its maximum texture size is via DISPLAY->GetMaxTextureSize().  This is a major bottleneck because GetMaxTextureSize(), a potentially slow call, gets hammered even though these values are not changing during runtime. Caching this value makes the game run much smoother. More importantly, on certain GPU's, such as AMD RX 6600 series, RageDisplay_OGL's GetMaxTextureSize() could be demonstrated to return an uninitialized variable on occasion. This seems to be because glGetIntegerv does not touch the variable if it fails. The reason this is being handled inside RageBitmapTexture rather than the OpenGL driver is because the frequent DISPLAY->GetMaxTextureSize() call is one of the largest bottlenecks in the texture loading process.