Update CPackSetup for WiX Toolset
Migrating from NSIS to WiX Toolset to generate a MSI installer for Windows platforms.

Notes:
WiX requests a version following the x.x.x.x format for the MSI installer, so we can need to append a .0 to the version number.
WiX also uses a GUID per product line for things like facilitating upgrades, etc.
The WixUI_InstallDir template provides a typical install wizard, which is nearly identical to our existing EXE installer as-is.