Disable macOS (Intel) CI runner
Takes the longest of any CI runner besides code analyzation by far (usually about 13 minutes,  sometimes longer). This uses a non trivial amount of electricity for something we honestly don't care a whole lot about. This is not proposing to stop support of macOS (Intel) and is not made with the intention of any follow-up commits.