Add SleepIfFocusLost to reduce CPU usage when unfocused
Introduces a method to yield the main thread a bit if focus is lost and calls it in EndFrame for D3D, GLES2, and OGL backends. Helps reduce CPU usage further when the game has lost focus, but maintains enough responsiveness for the game to run as expected and smoothly transition back into gameplay without stuttering when focus is regained.