macOS: Don't copy the entire songs folder into ITGmania.app when building.
The current code copies the entire contents of the /Songs folder within ITGmania.app itself.  So, if you have 20GB of packs when you compile the game, you end up with a 20GB sized executable. It's hard to imagine this was intended to do any more than maybe bundle the three default StepMania 5 songs, so I've removed it.