Refactor Club Fantastic integration in CMake
The current code has two problems. First, on macOS, it copies the entire contents of the /Songs folder within ITGmania.app itself. So, if you have 20GB of packs when you compile the game, you end up with a 20GB sized executable. Second, you can end up with two copies of Club Fantastic songs on macOS (one in the Songs folder and one in the .app bundle). This refactors Club Fantastic CMake integration to the top level so that all platforms can be built as expected and unwanted or duplicate song packs will not end up in the macOS builds.