Use integer time in ArrowEffects
ArrowEffects is one of the major callers of system time as a floating point. It is directly related to perceived visual sync matching with audio sync. Most of the game has been updated to directly use microseconds already. I have updated this to use the integer microseconds time value directly. Floating point calculations are made locally if needed from the microsecond value, so Lua gimmicks continue to work as expected. Const correctness is also improved in this commit.