Fix mixing issue which caused audio clipping on macOS & JACK
Clamping that was applied to the int16_t mix buffer read function also needed to be applied to the float mix buffer read function, which is used by macOS AudioUnit and JACK audio drivers. This was resulting in audio clipping which was only perceptible when using either of these drivers. This method replaces the manual loop with std::transform and a lambda for clamping values, improving code efficiency.