Build ffmpeg-w32 with GitHub Actions
Adapted the existing ffmpeg-w32 readme for Arch Linux to a GitHub Actions workflow for Ubuntu 24.04 - many thanks to natano for leaving the readme in the ffmpeg-w32 directory with the instructions needed to rebuild these libraries. The new workflow ensures generic x86_64 binaries are created, handles automatic assembly of the include directory, and enables the GPL 3 license flag.