Resolve CMake issue causing macOS builds to have duplicate copies of all resources
Currently, the macOS build creates a MACOSX_BUNDLE with resources inside it, but also installs those same resources to the ITGmania destination directory. 

This should hopefully clear up confusion for macOS users regarding whether the .app itself or the whole folder should be copied to Applications and ensure users do not attempt to keep custom content in the wrong place.